+++ lib/procodile/config.rb
@@ -27,6 +27,7 @@ module Procodile
       if @processes
         process_list.each do |name, command|
           if process = @processes[name]
+            process.removed = false
             # This command is already in our list. Add it.
             if process.command != command
               process.command = command
@@ -36,6 +37,15 @@ module Procodile
           else
             Procodile.log nil, 'system', "#{name} has been added to the Procfile. Adding it."
             @processes[name] = create_process(name, command, COLORS[@processes.size.divmod(COLORS.size)[1]])
+          end
+        end
+
+        removed_processes = @processes.keys - process_list.keys
+        removed_processes.each do |process_name|
+          if p = @processes[process_name]
+            p.removed = true
+            @processes.delete(process_name)
+            Procodile.log nil, 'system', "#{process_name} has been removed to the Procfile. It will be removed when it is stopped."
           end
         end
       end